start to pump pressure through consistently

+++ Bsp/cLcd.h
@@ -41,7 +41,7 @@ public:
   void info (std::string str, bool newLine = true);
 
   // touch
-  void press (int pressCount, int x, int y, int z, int xinc, int yinc);
+  void press (int pressCount, int16_t x, int16_t y, uint16_t z, int16_t xinc, int16_t yinc);
 
   void startRender();
   void renderCursor (uint32_t colour, int16_t x, int16_t y, int16_t z);
